+++ go.mod
@@ -3,7 +3,7 @@ module github.com/ElrondNetwork/elrond-go
 go 1.13
 
 require (
-	github.com/ElrondNetwork/arwen-wasm-vm v0.4.4-0.20201103182403-cdba621b66f8
+	github.com/ElrondNetwork/arwen-wasm-vm v0.4.4-0.20201103193320-060f1d5e90d3
 	github.com/ElrondNetwork/concurrent-map v0.1.3
 	github.com/ElrondNetwork/elrond-go-logger v1.0.4
 	github.com/ElrondNetwork/elrond-vm-common v0.3.1
